backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==1.10.13
python-dotenv==1.0.0
motor==3.7.1
pymongo==4.13.0
starlette==0.27.0
typing_extensions==4.13.2
anyio==3.7.1